+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/Store.js
@@ -1,0 +1,34 @@
+
+Ext.define('Fiddle.Records', {
+    extend: 'Ext.data.Store',
+    config: {
+        autoDestroy: true,
+        proxy: {
+            type: 'ajax',
+            url: 'data.json',
+            reader: {
+                type: 'json',
+                rootProperty: 'data',
+                totalProperty: 'total'
+            }
+        },
+        remoteSort: false,
+        sortInfo: {
+            field: 'name',
+            direction: 'DESC'
+        },
+        pageSize: 50,
+        fields: ['title', 'link', 'desc'],
+        autoLoad: true
+    }
+}, function () {
+    Ext.create('Fiddle.Records', {
+        storeId: 'records'
+    });
+    Ext.create('Ext.data.ChainedStore', {
+        source: 'records',
+        storeId: 'viewrecords'
+    });
+});
+
+