+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src_bck/meta.js
@@ -1,0 +1,9 @@
+var meta = {
+    fiddleHeader: 'Ext Itunes',
+    fiddleSubHeader: 'Open Source Itunes <i>like</i> UI' +
+    '<br />'
+};
+
+var fiddleStore = null;
+
+