+++ fiddles/nativeScript/fiddle-0008-FifteenPuzzle/fifteenpuzzle/platforms/android/src/main/assets/app/view/level-two/level-two.js
@@ -1,0 +1,10 @@
+var createViewModel = require("../../shared/main-view-model").createViewModel;
+var frame = require('ui/frame');
+
+function onButtonClick(args) {
+  console.log('onButtonClick');
+  console.log(args);
+}
+
+exports.onButtonClick = onButtonClick;
+